requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
pydantic
python-dotenv
requests
tensorflow
numpy
pandas
scikit-learn
